cc: move DEFAULT_SERVER_CONFIG_PATH to consts.py
